Tools Funcionales
Se realizaron algunos cambios en citas y posología para evitar errores y se implementaron las tools de búsqueda

+++ asistentesacyl/domain/src/main/java/es/upsa/tfg/domain/entities/Posologia.java
@@ -16,6 +16,6 @@ public class Posologia
     String id_paciente;
     String id_medicina;
     int dosis;
-    double frecuencia;
+    int frecuencia;
     String unidad;
 }
